nas_webserver/requirements.txt
Flask
flask-mysqldb
psutil
python-dotenv
werkzeug
